refactor: 불필요한 뎁스 제거를 위해 Deck 제거
