+++ src/test/java/blackjack/model/deck/HandDeckTest.java
@@ -1,6 +1,8 @@
 package blackjack.model.deck;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.junit.jupiter.api.Assertions.assertAll;
 
 import blackjack.model.card.Card;
 import blackjack.model.card.CardNumber;
@@ -13,7 +15,7 @@ class HandDeckTest {
 
     HandDeck handDeck = new HandDeck();
 
-    @DisplayName("핸드덱에 카드를 추가한다.")
+    @DisplayName("덱에 카드를 추가한다.")
     @Test
     void add() {
         //given
@@ -27,17 +29,65 @@ class HandDeckTest {
         assertThat(cards).containsExactly(card);
     }
 
-    @DisplayName("핸드덱에 카드를 추가할때 스코어를 계산한다.")
+    @DisplayName("덱에 카드를 추가할 때 동일한 카드가 있는 경우 예외를 발생시킨다.")
     @Test
-    void calculateDeckScore() {
+    void add_duplicateCard() {
         //given
-        Card card = new Card(CardPattern.CLOVER, CardNumber.EIGHT);
+        Card card1 = new Card(CardPattern.CLOVER, CardNumber.ACE);
+        Card card2 = new Card(CardPattern.CLOVER, CardNumber.ACE);
+
+        handDeck.add(card1);
+        //when, then
+        assertThatThrownBy(() -> handDeck.add(card2))
+                .isInstanceOf(IllegalArgumentException.class);
+    }
+
+    @DisplayName("덱에 있는 모든 카드의 합을 계산한다.")
+    @Test
+    void calculateCardScore() {
+        //given
+        Card card1 = new Card(CardPattern.CLOVER, CardNumber.EIGHT);
+        Card card2 = new Card(CardPattern.SPADE, CardNumber.EIGHT);
 
         //when
-        handDeck.add(card);
-        int score = handDeck.score();
+        handDeck.add(card1);
+        handDeck.add(card2);
 
         //then
-        assertThat(score).isEqualTo(CardNumber.EIGHT.getNumber());
+        assertThat(handDeck.calculateCardScore()).isEqualTo(16);
+    }
+
+    @DisplayName("덱에 11값으로 설정된 Ace 카드의 갯수를 반환한다.")
+    @Test
+    void countElevenAce() {
+        //given
+        Card card1 = new Card(CardPattern.CLOVER, CardNumber.ACE);
+        Card card2 = new Card(CardPattern.SPADE, CardNumber.ACE);
+
+        //when
+        handDeck.add(card1);
+        handDeck.add(card2);
+
+        //then
+        assertThat(handDeck.countElevenAce()).isEqualTo(2);
+    }
+
+    @DisplayName("덱에 11값으로 설정된 Ace 중 제일 첫번쨰 Ace의 값을 1로 변경한다.")
+    @Test
+    void switchAceValueInRow() {
+        //given
+        Card card1 = new Card(CardPattern.CLOVER, CardNumber.ACE);
+        Card card2 = new Card(CardPattern.SPADE, CardNumber.ACE);
+
+        //when
+        handDeck.add(card1);
+        handDeck.add(card2);
+        handDeck.switchAceValueInRow();
+
+        //then
+        assertAll(
+                () -> assertThat(card1.getScore()).isEqualTo(1),
+                () -> assertThat(card2.getScore()).isEqualTo(11)
+        );
     }
 }
